+++ src/WebApiModels/MapHandling/WebMercatorProjection.cs
@@ -43,7 +43,7 @@ namespace VSS.Productivity3D.WebApi.Models.MapHandling
       return new Point(lat, lng);
     }
 
-    public static Point LatLngToTile(Point latLng, int numTiles)
+    public static Point LatLngToTile(Point latLng, long numTiles)
     {
       return PixelToTile(LatLngToPixel(latLng, numTiles));
     }
@@ -66,21 +66,21 @@ namespace VSS.Productivity3D.WebApi.Models.MapHandling
       };
     }
 
-    public static Point LatLngToPixel(Point latLng, int numTiles)
+    public static Point LatLngToPixel(Point latLng, long numTiles)
     {
       Point worldPt = FromLatLngToPoint(latLng);
       Point pixelPt = WorldToPixel(worldPt, numTiles);
       return pixelPt;
     }
 
-    public static Point PixelToLatLng(Point pixelPt, int numTiles)
+    public static Point PixelToLatLng(Point pixelPt, long numTiles)
     {
       Point worldPt = PixelToWorld(pixelPt, numTiles);
       Point latLng = FromPointToLatLng(worldPt);
       return latLng;
     }
 
-    public static Point WorldToPixel(Point worldPt, int numTiles)
+    public static Point WorldToPixel(Point worldPt, long numTiles)
     {
       return new Point
       {
@@ -89,7 +89,7 @@ namespace VSS.Productivity3D.WebApi.Models.MapHandling
       };
     }
 
-    public static Point PixelToWorld(Point pixelPt, int numTiles)
+    public static Point PixelToWorld(Point pixelPt, long numTiles)
     {
       return new Point
       {
@@ -107,7 +107,7 @@ namespace VSS.Productivity3D.WebApi.Models.MapHandling
     /// <param name="longitude">longitude in radians</param>
     /// <param name="numTiles">number of tiles (calcuated from zoom level)</param>
     /// <returns>x tile coordinate</returns>
-    public static int LongitudeToTile(double longitude, int numTiles)
+    public static int LongitudeToTile(double longitude, long numTiles)
     {
       var columnIndex = longitude;
       var columnNormalized = (1.0 + columnIndex / Math.PI) / 2.0;
@@ -122,7 +122,7 @@ namespace VSS.Productivity3D.WebApi.Models.MapHandling
     /// <param name="latitude">latitude in radians</param>
     /// <param name="numTiles">number of tiles (calcuated from zoom level)</param>
     /// <returns>y tile coordinate</returns>
-    public static int LatitudeToTile(double latitude, int numTiles)
+    public static int LatitudeToTile(double latitude, long numTiles)
     {
       var rowIndex = Math.Log(Math.Tan(latitude) + (1.0 / Math.Cos(latitude)));
       var rowNormalized = (1.0 - rowIndex / Math.PI) / 2.0;
